Add util function to convert PerfResults to format.Format.
The encoder outputs the json format that conforms to [1]. It places
the trace key in the common key for each sample set, and it generates
several aggregated values for each measurement. The generated Json will
be used by the ingestor.

Note, the benchmark (encoder vs marshaler) is done in [2]. CPU is better
at a small cost on RAM.

[1]: perf/go/ingest/format/formatSchema.json
[2]: https://skia-review.googlesource.com/c/buildbot/+/844557/11

Bug: b/334901479
Change-Id: I36571f96625c183a1694c2c2b9e0b93565833d13
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/844557
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>