[pinpoint] retain swappability of catapult bisect workflow
The response object was updated to a new struct object that contains
the datastore response, but this makes the response type different from
internal.Bisect and makes this workflow less interchangeable with it.

The internal.BisectExecution object currently returns a good amount
of additional data to generate the information for Catapult. However,
it being under the internal package doesn't allow it to be exposed.
Anomaly workflows need access to Culprits.

To address both, response type is updated to pinpoint_proto, which
contains just the information about culprit while being accessible
outside of internal. Once the backwards compatibility is no longer
needed, the response object of internal.BisectWorkflow can be
reverted to the pinpoint_proto and the catapult package can be
deleted.

Bug: b/327032048
Change-Id: I2b88c9658a3db285dfb44a19f53afcd779d05f2c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/848487
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>