Fix up the Revision Info UI
There are cases where anomalies are created on the parent traces (eg: on subtest_1 instead of subtest_2). With filter_parent_traces set to true in the config, these get filtered out in the revision info url renderings and thus it appears as if there is no anomaly in the charts. Therefore this CL disables parent trace filtering for these urls.

- Add option to disable filtering parent traces in explore.
- Set this option to true when creating urls for revision info.
- Also update the way the revision infos are grouped so that all anomalies are visible and not only till the test param.

Change-Id: Ia925cc3f41483f3d5e1c89b8f23066f54500c3c3
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/848122
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Eduardo Yap <eduardoyap@google.com>