[pinpoint] address zeros in stats analysis for pairwise compare
The wilcoxon signed rank test by default takes the log transform of
the input data. However, it is common for pinpoint results to generate
zeros, which will throw error in logarithm functions.
When there are zeros in the input arrays, we use the normalized
transform instead

Bug: b/337946034
Change-Id: If8e17369d00008603e646be0bc7bba9a81447773
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/850366
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>