[pinpoint] create CulpritFinderWorkflow aka sandwich verification
The CulpritFinderWorkflow runs the PairwiseWorkflow,
CatapultBisectWorkflow, and then N PairwiseWorkflow, where N is the
number of candidate culprits. It is hosted in the catapult/ directory
since it relies on the CatapultBisectWorkflow.

The workflow is unit tested but not yet tested on temporal. Since
the pairwise workflow is still WIP, it makes more sense to confirm
the pairwise workflow works then revisit the functionality of the
sandwich workflow.

Bug: b/318862550
Change-Id: I2bc97602ec2ac5a11373b27a8cfd145a6b69e465
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/851734
Reviewed-by: Hao Wu <haowoo@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>