Integrate test-picker-sk to MultiGraph.
Feature is guarded on the useTestPicker URL param and by the inclusion of the query_config.include_params flag in the instance config.

If both of the above conditions are met, the Test Picker is enabled and the old Add Graph button is disabled. Furthermore, a user can click on a trace and a new "Query Highlighted" button will appear. Clicking this button will populate the Test Picker with the highlighted trace's keys.

Bug: chromium:340936052
Bug: chromium:340936786
Bug: chromium:340936518

Change-Id: I59d3f97a9636118e6483e67412b81b4cd863c4a2
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/859542
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>