[pinpoint] handle all values equal edge case in ComparePairwise
When every element of valuesA is identical and every element of valuesB
is identical, the wilcoxon signed rank test is unable to estimate a
confidence interval. However, the underlying data is still useful. We
nudge one of the values by an insignificant amount to unblock the
analysis and return a valid confidence interval.

This edge case is not supported in catapult and will return an error,
blocking bisection.

Bug: b/332612913
Change-Id: I90e1397cd2e5dd04a0a22420caa311d13e7945af
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/862501
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>