[pinpoint] do not search waterfall builders when bisecting DEPS commits or commits with gerrit patches
Prior to this CL, when bisection finds a commit with a DEPS roll, it
will recycle a build without the DEPS roll because bisection will
look in the waterfall pool for the chromium/src build. This CL ensures
searchBuild skips the waterfall pool if the build has DEPS or patches.

While bisection currently does not use gerrit patches, it is useful
for future pinpoint jobs to skip searching in the waterfall pool if a patch is submitted

Bug: b/346362465
Change-Id: Id4bd3ef07467271c2992d13ef529afb491478826
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/864322
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>