[chromperf] add caching in psrefresh to reduce query latency
This is the plan b of using Redis as the cache. Due to the blockers on network accesses, the caching logic is implemented in frontend's in-memory cache first. When Redis is ready, we can move the logic there. The ideas of caching the top 2 levels of query parameters will be mostly the same.

Change-Id: I118fbb302c711c9ca0b3cb9aefa4a3cc53a35a8e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/865801
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>