Refactor default fetching out of explore-simple-sk and make test-picker-sk use defaults.
Previously, every time an ExploreSimpleSk element was initialized, it would fetch default configs from the backend. This is changed to be done by the MultiGraph element. This way there are less calls and we can pass defaults to the TestPicker element as well.

TestPicker element now uses defaults to modify the query it sends to /_/nextParamList/.

Bug: chromium:340936052

Change-Id: I631c22de56b98d6f251c62f3acb8bc4e7571e0fa
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/864914
Commit-Queue: Eduardo Yap <eduardoyap@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>