Use SideEffect to build randomized string.
In the replay, the UUID will generate a different string, causing
non-determistic error. SideEffect should be used for cases like external
dependencies.

Also note the versioning here is not needed, because once the worker is
restarted, the old workflow will not be able to pick up due to this
non-determisitism.

Change-Id: Icd7d53de7e9e714e4955d5b5ff40b3eefd4e65c1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/868096
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>