[perf] expose git repo config url to JS
The git repo config contains the git url which can be used to call
gitiles APIs. This is nested within each instance's config file.

This change exposes that url to JS running in the browser so that
it can build the links to gitiles.

Bug: b/338440689
Change-Id: Iaf15f8ce8caa1d6b99ab5def595ff45819bfb08a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/869677
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>