[pinpoint] skip functional test results during catapult writeback
Functional test results doesn't set commit information, so the
commitpairvalues are nil for lower and higher. This threw a
nil pointer.

Catapult writeback is only interested in performance based data,
so functional comparisons can be ignored. This change checks
ResultType before adding it to the map. As a result, the Performance
and Functional constants are made public (upper case).

Change-Id: Ie720380c90ffc8d670a70ca438b330f4aeb47a5c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/870366
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>