[pinpoint] fix non-deterministic fetch free bots unit test
FetchFreeBots gets a list of swarming dimensions, converts
the dimensions to a map, and then converts dims to a StringPair array.
The map to StringPair conversion done by StringMapToTaskDimensions
generates the dimensions in a non-deterministic order, occasionally
failing the unit test TestFetchFreeBots_ForBuilder_ReturnsFreeBots.
The mock call for ListBots expects the dimensions to be fixed.

We remove any reliance on StringMapToTaskDimensions in FetchFreeBots.

Bug: b/322244008
Change-Id: I13af5aa0cf349eecf498101742a457ad8020f3a1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/872112
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>