[pinpoint] monitor pairwise task ordering and task interception
Pairwise tasks need to run in order and without interruption to ensure
they uphold all of the assumptions used by the wilcoxon signed rank
test. Swarming should be able to uphold task pair ordering but cannot
guarantee task interception, as it is possible for two requests to
lay on top of each other and cause interceptions.

Because of how crucial this assumption is, we add monitoring for this
step so we can determine how frequent the assumptions are violated
and if they are worth fixing.

Bug: b/344985109
Change-Id: Ib32edc869581054543aa8f68a520b44d9dd5f269
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/874836
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>