[pinpoint] verify culprit against previous commit instead of starting commit in culprit finder workflow
In the sandwich verification workflow between commits A vs B we run:
- regression verification (A vs B)
- bisect (A vs B)
- culprit verification (C-1 vs C)

The culprit verification step ideally compares the culprit against
the commit prior to the culprit (C-1 vs C). However, bisection only
returns a list of culprits. Since culprit verification does not have
access to the commit prior, we do a suboptimal comparison A vs C. This
comparison works but it would be better to do C-1 vs C.

This CL adds DetailedCulprits to the `BisectionExecution`, which
contains the culprit and the commit prior to enable C-1 vs C
comparisons.

Bug: b/340220164
Change-Id: I30cca9c7cf11b15f228b1cce68a4249d4bc48b49
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/870753
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>