[pinpoint] culprit finder passes magnitude from regression verification to bisect
Anomaly detection calculate a comparison magnitude which estimates
how large the expected regression is going to be. However, with high
variance data, it is easy for this magnitude to be way off from the
real regression.

Regression verification takes 30 samples of each commit, which is
enough sample size to calculate a new comparison magnitude that can
be passed to bisection.

Bug: b/340884347
Change-Id: I98277f1f676bc1dd284c3b7c2639b3487b5f3c9e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/876456
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Auto-Submit: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>