Fix Ansible scripts for updating jumphosts etc
ansible-playbook ./switchboard/jumphosts.yml
was failing in two different ways.
1) We couldn't target rack1 because of an error like:

TASK [Gathering Facts] *******************************************************************
fatal: [rack1]: FAILED! => changed=false
  ansible_facts: {}
  failed_modules:
    ansible.legacy.setup:
      ansible_facts:
        discovered_interpreter_python: /usr/bin/python3
      failed: true
      msg: 'ansible-core requires a minimum of Python2 version 2.7
   or Python3 version 3.6. Current version: 3.5.2 (default, Jan 26 2021, 13:30:48) [GCC 5.4.0 20160609]'
  msg: |-
    The following modules failed to execute: ansible.legacy.setup

The jumphost on rack1 is running an embarrassingly old version of
Ubuntu, so it only has Python 2.7.12 and 3.5. We can't update to 3.6+
(I tried, the old ppa deadsnakes dropped support for Xenial 2 years
ago) so instead we use the ansible_python_interpreter [1] to tell Ansible
ignore the Python3 version which is too old and use the supported
but old Python 2.7.

2) The script was failing partway through on all racks because
a step in copy_authorized_keys was failing. In particular, the
part that copies the jumphost's public key to the target machine
(in this case, also the jumphost) was failing because the variable
copy_authorized_keys__jumphost was not set, so the command was invalid.
The intention appears to be to skip these steps when the variable
is not set, but that doesn't work (maybe due to an Ansible update).
So, I used a sentinel value "skip" to make it be set but also
skipped as per the `when` clause.

Now the scripts run to completion on all jumphosts (after running
the command to get breakglass access for the secrets).

[1] https://docs.ansible.com/ansible/latest/reference_appendices/python_3_support.html

Bug: b/351886693
Change-Id: I4c480272fed07829b6878ea7d97fd7dfa60638ce
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/878536
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>