[pinpoint] always append CommitPairValues in CompareActivity result
The catapult bisect UI parser expects that certain values will always
be there during the UI write. However, functional analysis tends to
make shortcuts in the response to try and minimize the amount of data
that is returned, which causes the parser to hit nil pointers.

This CL always returns the CommitPairValues to try and avoid these
issues in the future

Bug: b/354035478
Change-Id: Ie5adb60651d261ac444a2c2d29b1619e94a8a30a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/879308
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>