[tmm] Add in code to detect types of EVK using PyOCD
This assumes pyocd is installed on the host machine. If we need
more things to be reported through Swarming, we can enhance
this, including rebooting the device, which is currently handled [1].

Because devices can have different output via pyocd, we can't rely
on them for getting a device type. Therefore, we will put a file
on the host to say what it is and use pyocd to verify that the
device is attached.

I chose not to handle this via the UI like we do for ChromeOS because
I didn't want to pipe the UI through for that, given that we'll only
have a few test boards. We can revisit if we have enough engineering
time.

Example contents of /etc/skia/pyocd_config.json
{
  "device_type": "ABC12345-EVK"
}
This is JSON5, so we can add comments as necessary

[1] https://source.corp.google.com/h/skia/codesearch/+/main:eskia/infra/bots/test.py;l=542;drc=b5bce4a2096da17c35fbf8a3f9d08f8d1c262463

Bug: b/40044966
Change-Id: I573b43c85561bdd503b48dbe0b92967a81551476
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/877521
Reviewed-by: George Wright <wrightgeorge@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>