[pinpoint] Create generic build client interface and chrome impl
The current implementation of Pinpoint has close dependencies to
Chrome and its infra stack, relying on Buildbucket, Swarming,
RBE CAS and Gitiles.

This is the first step in abstracting away general functionality
to support projects that onboard to Pinpoint, that stray away from
the LUCI stack.

This change introduces a build client that defines a generic
interface that can be utilized by a generic build workflow.
The underlying client, build_chrome for Chrome, implements
the Chrome specific logic.

The client introduced here is not yet integrated with any workflow.

Bug: chromium:353273333
Change-Id: Ia1af6e9a0bb5efa33b73dee977165e38f5d1ba66
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/879999
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>