[pinpoint] integrate generic build client into build_chrome
This change takes the generic build client and integrates into the
build_chrome workflow.

Note that there are still a lot of ties to "Chrome" in this change.
This change focuses on ensuring that there are no functional
regressions, and there will be follow up changes to remove "chrome"
specific language.

Bug: chromium:353273333
Change-Id: I1097804426847718e52d270f93efb41e08fefd78
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/880737
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>