[pinpont] Make build workflow fully agnostic
The BuildChromeWorkflow is renamed to BuildWorkflow with this change.

The BuildWorkflow instantiates a BuildClient, which will depend
on the project that the build workflow is running for. This will
default to Chromium, and will be Chromium for the time being.

The BuildClient is added to the context, such that all sub activities
or workflows can re-use the same client instead of instantiating its
own client. Should the client be unavailable, each activity/workflow
should create a client for execution.

Bug: chromium:353273333
Change-Id: I4f675fa5cc724705039ce77095252c66c2fbb520
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/880727
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>