[perf] add DeleteShortcut to Shortcuts CockroachDB table
In the skia-perf cluster, the shortcuts table occupies about 50% of
the skia database's total memory. There is no TTL for this data so it
will persist indefinitely. In the short term, we need to delete
some of these shortcuts to avoid reaching full capacity.

Bug: b/354762705
Change-Id: I3e16a95ca8e9b4cfb19d6442b8494a622a60aa48
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/882501
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>