[task scheduler] Set Finished time in Buildbucket job cancellation
I got a couple of alerts indicating that we hadn't sent notifications to
Buildbucket for a handful of Jobs for over 5 minutes. Investigation
revealed that these Jobs were cancelled by Buildbucket itself due to a
new patchset being uploaded to the CL and the jobs becoming outdated.

It turns out that there's a race condition in our metrics in which we
pulled the build from Buildbucket while it was still "started" and then
pulled the Job from the DB after cancellation took place, so we wrote
the metric data to indicate the lag time between the job finishing and
the build being marked finished, even though by the time the metric was
reported both were marked finished.

However, it also turns out that the Finished timestamp on the job wasn't
set, and with the timestamp being "0001-01-01T00:00:00Z", the reported
lag time was enormous instead of very small and thus triggered the
alert.

This CL fixes the Finished timestamp to be properly set when jobs are
cancelled by Buildbucket, which should prevent these noisy alerts. I
also added a warning indicating when we're inserting a job into the DB
with a status indicating that it's finished but no Finished timestamp. I
also noticed that the job's StatusDetails weren't being reported to the
UI due to an oversight in rpc.go.

Change-Id: I7c06379442a9342c4be9173a1a9f1a8555ea8cca
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/885042
Auto-Submit: Eric Boren <borenet@google.com>
Commit-Queue: Eric Boren <borenet@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>