[chromeperf] add pinpoint activity to trigger find culprit workflow
Instead of invoking the find culprit as a child workflow and wait, now we call the pinpoint endpoint, ScheduleCulpritFinder, to asynchronously start the workflow.

Note that the job id is now generated by the client and passed to the find culprit workflow. It is no longer returned from the workflow.

Change-Id: Ibf9e0e67ab86f1c17201a7eb368971fca12ae539
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/885377
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>