[chromeperf] hardcode the pinpoint queue name for now for integration
The config value is not available and it is not supposed to be directly read from the worker. I'm making this change to pass the hardcoded value for now. Will pass the value as part of the request.

Bug: b/353979260
Change-Id: I3b4cc1acd730d81ddf8de212b4f3a41e0204cb2d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/886522
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>