[chromeperf] filter invalid aggregation values and return
Noticed the the 'stat' value can be 'error' in the anomaly. While we try to use the 'stat' value to be the aggregation method, we should only allow the valid values before we create the bisection child workflow.

The allowed aggregation methods are in:
https://github.com/google/skia-buildbot/blob/d7d6072b1cbad96466ed51959a37502176ef08b2/pinpoint/go/read_values/read_values.go#L16

Bug: b/357619045
Change-Id: Ice59346227bf7367ae3fe078b8e88a9fa31a9916
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/887911
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>