[chromeperf] update the chart name and use the suffix of the chart for stat when triggering bisect
This CL update the chart and stat value for bisect job. The logic is from the legacy where if the chart name has a suffix of one of the six:
  "avg", "count", "max", "min", "std", "sum",

the suffix will be used as the 'stat' value for the new bisect job.

If the 'stat' is 'avg', the skia bisect will convert it to 'mean'.

e.g.;
if the chart name is "metric_a_on_b_avg", we will have "metric_a_on_b" as chart and "avg" as stat. Skia Bisect will convert "avg" to "mean".
if the chart name is "metric_a_on_b_special", we will have "metric_a_on_b_special" as chart and "" as stat. Skia bisect will use the default aggregation method.

Bug: b/357619045
Change-Id: I377512d83907a465043dad4c179c87d65390f1bc
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/889899
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>