[chromeperf] add culprit service url and anomaly group id to find culprit request
Those values are needed as part of the anomaly grouping workflow: when pinpoint finds the culprit(s), if the job is initiated as auto-bisect with an anomaly group ID, we should invoke the culprit service workflow to process the culprit(s) accordingly.

Please note that the culprits are not returned yet. Will do it after sync between the culprit proto definitions between the culprit service and the bisection. Currently they are different.

Bug: b/357619045
Change-Id: I2a664019b643530cc2f15fd1bdc5f6bfcd20e7e0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/886247
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>