[goldctl] Add context to httpclient abstraction
goldctl does not seem to be able to retry a request that times out after taking
longer than 30 seconds due to not being able to abort Get or Post early
and retry. By adding in Contexts, we should be able to support this and
actually retry. In a follow-up CL, I'll add such support.

Bug: b/354748051
Change-Id: Iee0c99ef5aff45586698982b54eb59b7c4bd8402
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/890396
Reviewed-by: Eric Boren <borenet@google.com>