[chromeperf] use new notify type for culprit processing
Updated the definition of 'notifier' in the context of culprit detection.

The term 'notifier' is overloaded. The existing logic refer to the 'notifier' in the context of anomaly detection which is now set to 'anomalygroup' for chromium internal instance. It should refer to the 'notifier' in the context of culprit reporting.

The CL separate the definitions.

Bug: b/321082416
Change-Id: Id742c9a727179fd97fec588fea3ee7da20f2b35c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/890917
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>