[chromeperf] Update ProcessCulpritParam to let bisect pass back pinpoint commit
Currently the culprit finding workflow does not generate all the info needed for culprit processing, and we should not impose it as part of culprit finding workflow.

Instead, it provides the repository url and the commit hash which should be sufficient for to generate the commit on culprit proto within the culprit processing workflow.

This CL updates the data types to let the find culprit workflow returns the commit on pinpoint proto.

Bug: b/321082416
Change-Id: If0183f330693e5a05c90047f1409e7ab7367b694
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/890764
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>