Add script to create Docker image to build Skia and run fiddler
See linked bug for overall design and plan.

This is inspired by cipd-base [1] with some modifications based
on skia-build-tools [2] and the existing fiddler container [3].

As a follow-up, I'll create a Louhi flow [4] to make this image, kick
off a flow and then add a new skia_app_container that uses it for
fiddler-base.

[1] https://github.com/google/skia-buildbot/blob/504f49204b2117774e29927f14ef2a04cc41994a/kube/build_base_cipd_release
[2] https://github.com/google/skia/blob/04ce2e2bfc3536dfac23224b116b4473d32d436d/docker/skia-build-tools/Dockerfile
[3] https://github.com/google/skia-buildbot/blob/504f49204b2117774e29927f14ef2a04cc41994a/fiddlek/BUILD.bazel#L87-L96
[4] http://go/louhicl/158081

Bug: b/361559681
Change-Id: Ibb8201aae1b475164bbd69b3ae2e6f0154c8eac9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/892538
Reviewed-by: Eric Boren <borenet@google.com>