[fiddler] Build security wrapper into fiddler-build-skia image
Building the fiddler-base image on Louhi doesn't seem to like to shell
out to Docker. That is probably unnecessary since the goal of that
earlier behavior was to make sure the fiddle_secwrap was built in
a way that was compatible with the fiddler image. We can do that
more cleanly by just building it on *on* the fiddler image (such that
glibc and friends match).

I chose to put this in fiddler-build-skia because the security
wrapper changes very infrequently, so we don't need to build this
regularly like building the fiddler backend.

Bug: b/361559681
Change-Id: Ibc93621e3ec2be02036a78a8ebe6a880b45951b8
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/892178
Commit-Queue: Eric Boren <borenet@google.com>
Reviewed-by: Eric Boren <borenet@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>