[perf] update plot builder to support anomaly information
DataPoint is updated to include a boolean field to determine
whether the given point is an anomaly.

This should likely be updated in the future to contain actual
details of the actual anomaly, but for the sake of integrating
Google Charts, this should be sufficient for now.

Bug: b/362831653
Change-Id: I3a4603196eb29bcf1ccc2b3f574e17017980f516
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/894520
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>