Make the close functionality consistent with other close buttons
Aligned the close (other action) buttons on the bottom right and put a cross icon on the top right for:
- Tooltip
- Legacy query dialog
- Bisect Dialog
- Chart (Removed "Remove all" button as the cross icon does the job)
- New/Edit Favorite dialog
- Algo dialog on the dry run page

Note: The "Remove all" button is working behind a flag. It is by default show but if the users want to hide it then add a query param show_remove_all=false.

Bug: b/361511141
Change-Id: Ica0e6b353c8c6842fb7b00fc07e6d40a527b415e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/895037
Commit-Queue: Vidit Chitkara <viditchitkara@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>