[perf] transition liveness handler to a safer query
Previous the liveness handler used the regression store's oldest
commit query as a simple and straightforward way to check if the
connection to cockroachDB was still live. We found out that that query
will fail when the table has no data to query from.

Switching instead to a query that is agnostic to whether there is
any existing data.

Bug: b/361129890
Change-Id: Iadc137ad89a8b3e596b50763ba8649fdafbebe3f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/896946
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>