[perf] change buttons to material design, agg chart options
This change merges all chart options that were checkboxes into
a settings button with a dialog. The check boxes are changed
to switches to better indiciate whether something is on or off.

All buttons are changed to material design with this change
for a more consistent ui.

Change-Id: I08950f7e9c6fe7917377f5b7a93d4e7834d881c0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/899130
Reviewed-by: Hao Wu <haowoo@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>