Use TS5 and enable decorators.
Decorators were introduced from TS 5.0. Its syntax is not foreign and
can bring the syntax sugar to make the code clean.

https://devblogs.microsoft.com/typescript/announcing-typescript-5-0/#decorators

This also introduces a breakage in @types/codemirror that needs to upgrade. We are already on the latest codemirror, but the type is wrong. It doesn't harm because the extra properties are ignored at runtime. Other upgrades due to TS5 compatibilities:
1. eslint needs to upgrade because newer TS versions;
2. Nodejs is upgraded from 16 to 18, due to eslint dependency;
3. linter plugins changes due to eslint upgrades;
4. express needs to upgrade because puppeteer uses it.

Those upgrades are for tooling, and shouldn't affect production code.

Bug: b/361365957
Change-Id: I80601c419a0a9c2b084c05f9ca976bf4ced06a4a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/899033
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>