[perf] address markdown links in benchmark configs
In some perf front end instances, data uploaded to ingest-file-link-sk
will include markdown links. We can either correct the link upstream
as data is ingested or downstream in the front end module. Both fixes
are hacky so we apply the fix in the front end where the code is
version controlled. The ingestion dataflow job is not version
controlled.

Bug: b/361496903
Change-Id: I214914494605f440e4cb7b7d5a33d1f640384b0f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/898547
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>