Create "New Bug" dialogue for untriaged anomalies.
If an anomaly is untriaged (bug Id of 0), we display a "New Bug" button on the tooltip. The user can click it to display a dialog. In this dialog, user can request backend to triage this anomaly for them. If a bug is successfully created, the user gets re-directed to this new bug.

Other changes:
- If bugId is -1, display "Ignored Alert" in tooltip, instead of -1.
- Chart tooltip close button style.

Bug: b/365007853
Bug: b/365007911
Change-Id: I905e58d6091c42983caced20ba70a968455991dc
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/899979
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>