[Triage Workflow] Invalidate cache when anomaly is updated.
Now that anomalies can be modified from Skia UI, we need to invalidate them in the cache. Cache keys consist of traceName, startCommit and endCommit. startCommmit and endCommit information is not clearly available when modifying an anomaly. Therefore, we cannot modify the cache directly. Instead, we use a map to keep track of which traceIds have been modified. All anomalies with that traceId will then be invalidated.

Not the most precise solution but should work without having to completely disable caching.

Bug: b/368364042
Change-Id: I840b67fe41e2f9f2358ec5d57d60392048e53991
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/902938
Commit-Queue: Eduardo Yap <eduardoyap@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>