Update tag in yml.
The Image tag will be replaced with the built image so they can be
exposed in the container environment or the cli flags. This helps the
runtime environment to find out where the build is coming from.

We need the `@tag` to be uniquely identifiable, the tag parser needs
to do some more work to identify the actual tag.

Change-Id: I92e93051df9e2520eb67c3bb5a3bf5905ea49ea3
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/907096
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Auto-Submit: Hao Wu <haowoo@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>