Fix the selection update.
When the first time the chart is loaded, the `selectedRange` doesn't get
updated so the values are not offsets based on the selection. This CL
moves the selection update into the single function to handle this when
there is a data update. More specifically, the `_dataframe` stores the
current range of the dataframe values, and `selectedRange` stores the
index of the current range. Those two are combined to load the actual
commit data from `dfRepo.dataframe`.

Bug: b/372038843
Change-Id: Idcf81bbdef880a197621acfa3c9adcbc1317d577
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/907617
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>