Increase timeout for task-scheduler repo syncing
We were observing the syncing to k8s-config repo taking just longer
than 15 minutes on skia-infra-corp (but not skia-infra-public).

Not sure why this is, but to get things moving, I'm doubling the
timeout and adding some more tracing so this type of timeout
is easier to deduce from the logs and error messages.

Also, some of the metrics might not have been synced when we called
fatal out of main.go, so I added in some conservative flushes to
capture those (e.g. the sync times)

Bug: b/372201307
Change-Id: Ib692bf0d5ec1233303bb7b44873070726dec7b7b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/907488
Reviewed-by: Ravi Mistry <rmistry@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>