[perf] move the successful popup toast for triage to explore-simple-sk
The toast element for triage messages are defined within new-bug-dialog.sk and existing-bug-dialog.sk. Those two dialogs are under triage-menu-sk, which will be hidden when the triage request successfully finished. As a result, the triage messages are hidden along with the trige-menu-sk, and not seen by end users.

This CL adds another toast on explore-simple. When the 'anomaly-changed' event is caught, other than redrawing, the toast message will show up if any 'detail' is sent with the event.

Bug: b/371634105
Change-Id: Id8a73eb07782572103dad7cb67f30f3e10745175
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/909121
Reviewed-by: Eduardo Yap <eduardoyap@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>