Refactor tooltip position calculation.
The logic to place the tooltip is moved in the tooltip self. Its
position is now computed relative to its parent as how it is laid out.
This helps to place the tooltip relative to any parent and then can be
moved around if needed.

Change-Id: I752f5224ec374affa13da717eb3102fc3365d638
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/909814
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>