Add panning to GoogleChart.
When the mouse is pressing on the chart, it triggers panning if there is
no other interactions. It emits two events, selection-changing and
selection-changed, as their names indicate. The selection-changed event
will be emitted when the mouse is up.

When the panning extends to the data range, it will trigger to load additional data on selection-changed.

Bug: b/362831653
Change-Id: I1aaf8547ed38871c7b9bc97d81e040518798f369
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/913027
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>