[perf] adjust tooltip moveTo function's vertical movement
Prior to this CL, the tooltip would show either above or below the
data point. This is fine when the tooltip is small. However, the
tooltip can be tall enough where if it is showing above the data point,
the tooltip cuts off at the top, making it useless.

This CL changes the tooltip's vertical placement and shifts it based
on how much the tooltip exceeds the viewport.

Bug: b/370804498
Change-Id: I4eb4ba834c034cf8f776f41f5fba5598975567b9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/913969
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>