[perf/android] remove the defaults being set for params
The default params are causing conflict. In multigraph, because of the
guided format, the test class for JetsnackScrollBenchmark returns no
results because the defaults are being applied. For single chart,
we have internal conflicts between what the user is providing and
the defaults being set on the instance config.

This removes the default configs for now to allow the queries to
go through.

Bug: b/376513834
Change-Id: Iefb65fbfd9f66aaa634049a0f8f64727efc33fdf
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/914881
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>